refactored password logic
